# hindsight-docs/static/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hindsight HTTP API",
    "description": "HTTP API for Hindsight",
    "contact": {
      "name": "Memory System"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/default/banks/{bank_id}/reflect": {
      "post": {
        "tags": [
          "Memory"
        ],
        "summary": "Reflect and generate answer",
        "description": "Reflect and formulate an answer using bank identity, world facts, and opinions.\n\nThis endpoint:\n1. Retrieves experience (conversations and events)\n2. Retrieves world facts relevant to the query\n3. Retrieves existing opinions (bank's perspectives)\n4. Uses LLM to formulate a contextual answer\n5. Extracts and stores any new opinions formed\n6. Returns plain text answer, the facts used, and new opinions",
        "operationId": "reflect",
        "parameters": [
          {
            "name": "bank_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bank Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReflectRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReflectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Budget": {
        "type": "string",
        "enum": [
          "low",
          "mid",
          "high"
        ],
        "title": "Budget",
        "description": "Budget levels for recall/reflect operations."
      },
...
      "CreatedMentalModel": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Mental model ID"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Human-readable name"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "description": "What this model tracks"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description"
        ],
        "title": "CreatedMentalModel",
        "description": "A mental model created during reflection."
      },
...
      "FactsIncludeOptions": {
        "properties": {},
        "type": "object",
        "title": "FactsIncludeOptions",
        "description": "Options for including facts (based_on) in reflect results."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReflectBasedOn": {
        "properties": {
          "memories": {
            "items": {
              "$ref": "#/components/schemas/ReflectFact"
            },
            "type": "array",
            "title": "Memories",
            "description": "Memory facts used to generate the response"
          },
          "mental_models": {
            "items": {
              "$ref": "#/components/schemas/ReflectMentalModel"
            },
            "type": "array",
            "title": "Mental Models",
            "description": "Mental models accessed during reflection"
          }
        },
        "type": "object",
        "title": "ReflectBasedOn",
        "description": "Evidence the response is based on: memories and mental models."
      },
      "ReflectFact": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          },
          "context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context"
          },
          "occurred_start": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occurred Start"
          },
          "occurred_end": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occurred End"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "ReflectFact",
        "description": "A fact used in think response.",
        "example": {
          "context": "healthcare discussion",
          "id": "123e4567-e89b-12d3-a456-426614174000",
          "occurred_end": "2024-01-15T10:30:00Z",
          "occurred_start": "2024-01-15T10:30:00Z",
          "text": "AI is used in healthcare",
          "type": "world"
        }
      },
      "ReflectIncludeOptions": {
        "properties": {
          "facts": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FactsIncludeOptions"
              },
              {
                "type": "null"
              }
            ],
            "description": "Include facts that the answer is based on. Set to {} to enable, null to disable (default: disabled)."
          },
          "tool_calls": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ToolCallsIncludeOptions"
              },
              {
                "type": "null"
              }
            ],
            "description": "Include tool calls trace. Set to {} for full trace (input+output), {output: false} for inputs only."
          }
        },
        "type": "object",
        "title": "ReflectIncludeOptions",
        "description": "Options for including additional data in reflect results."
      },
      "ReflectLLMCall": {
        "properties": {
          "scope": {
            "type": "string",
            "title": "Scope",
            "description": "Call scope: agent_1, agent_2, final, etc."
          },
          "duration_ms": {
            "type": "integer",
            "title": "Duration Ms",
            "description": "Execution time in milliseconds"
          }
        },
        "type": "object",
        "required": [
          "scope",
          "duration_ms"
        ],
        "title": "ReflectLLMCall",
        "description": "An LLM call made during reflect agent execution."
      },
      "ReflectMentalModel": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Mental model ID"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Mental model name"
          },
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Mental model type: entity, concept, event"
          },
          "subtype": {
            "type": "string",
            "title": "Subtype",
            "description": "Mental model subtype: structural, emergent, learned"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "description": "Brief description"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary",
            "description": "Full summary (when looked up in detail)"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "type",
          "subtype",
          "description"
        ],
        "title": "ReflectMentalModel",
        "description": "A mental model accessed during reflect."
      },
      "ReflectRequest": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "budget": {
            "$ref": "#/components/schemas/Budget",
            "default": "low"
          },
          "context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context",
            "description": "DEPRECATED: Additional context is now concatenated with the query. Pass context directly in the query field instead. If provided, it will be appended to the query for backward compatibility.",
            "deprecated": true
          },
          "max_tokens": {
            "type": "integer",
            "title": "Max Tokens",
            "description": "Maximum tokens for the response",
            "default": 4096
          },
          "include": {
            "$ref": "#/components/schemas/ReflectIncludeOptions",
            "description": "Options for including additional data (disabled by default)"
          },
          "response_schema": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Schema",
            "description": "Optional JSON Schema for structured output. When provided, the response will include a 'structured_output' field with the LLM response parsed according to this schema."
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags",
            "description": "Filter memories by tags during reflection. If not specified, all memories are considered."
          },
          "tags_match": {
            "type": "string",
            "enum": [
              "any",
              "all",
              "any_strict",
              "all_strict"
            ],
            "title": "Tags Match",
            "description": "How to match tags: 'any' (OR, includes untagged), 'all' (AND, includes untagged), 'any_strict' (OR, excludes untagged), 'all_strict' (AND, excludes untagged).",
            "default": "any"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "ReflectRequest",
        "description": "Request model for reflect endpoint.",
        "example": {
          "budget": "low",
          "include": {
            "facts": {}
          },
          "max_tokens": 4096,
          "query": "What do you think about artificial intelligence?",
          "response_schema": {
            "properties": {
              "summary": {
                "type": "string"
              },
              "key_points": {
                "items": {
                  "type": "string"
                },
                "type": "array"
              }
            },
            "required": [
              "summary",
              "key_points"
            ],
            "type": "object"
          },
          "tags": [
            "user_a"
          ],
          "tags_match": "any"
        }
      },
      "ReflectResponse": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "based_on": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReflectBasedOn"
              },
              {
                "type": "null"
              }
            ],
            "description": "Evidence used to generate the response. Only present when include.facts is set."
          },
          "structured_output": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Structured Output",
            "description": "Structured output parsed according to the request's response_schema. Only present when response_schema was provided in the request."
          },
          "usage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TokenUsage"
              },
              {
                "type": "null"
              }
            ],
            "description": "Token usage metrics for LLM calls during reflection."
          },
          "trace": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReflectTrace"
              },
              {
                "type": "null"
              }
            ],
            "description": "Execution trace of tool and LLM calls. Only present when include.tool_calls is set."
          },
          "mental_models_created": {
            "items": {
              "$ref": "#/components/schemas/CreatedMentalModel"
            },
            "type": "array",
            "title": "Mental Models Created",
            "description": "Mental models created during this reflection (via the learn tool)."
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "ReflectResponse",
        "description": "Response model for think endpoint.",
        "example": {
          "based_on": {
            "memories": [
              {
                "id": "123",
                "text": "AI is used in healthcare",
                "type": "world"
              },
              {
                "id": "456",
                "text": "I discussed AI applications last week",
                "type": "experience"
              }
            ],
            "mental_models": [
              {
                "description": "Understanding of AI capabilities",
                "id": "mm-1",
                "name": "AI Technology",
                "subtype": "structural",
                "type": "concept"
              }
            ]
          },
          "mental_models_created": [
            {
              "description": "Track AI-related decisions and plans",
              "id": "mm-new-1",
              "name": "AI Strategy"
            }
          ],
          "structured_output": {
            "key_points": [
              "Used in healthcare",
              "Discussed recently"
            ],
            "summary": "AI is transformative"
          },
          "text": "Based on my understanding, AI is a transformative technology...",
          "trace": {
            "llm_calls": [
              {
                "duration_ms": 1200,
                "scope": "agent_1"
              }
            ],
            "tool_calls": [
              {
                "duration_ms": 150,
                "input": {
                  "query": "AI"
                },
                "tool": "recall"
              }
            ]
          },
          "usage": {
            "input_tokens": 1500,
            "output_tokens": 500,
            "total_tokens": 2000
          }
        }
      },
      "ReflectToolCall": {
        "properties": {
          "tool": {
            "type": "string",
            "title": "Tool",
            "description": "Tool name: lookup, recall, learn, expand"
          },
          "input": {
            "additionalProperties": true,
            "type": "object",
            "title": "Input",
            "description": "Tool input parameters"
          },
          "output": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output",
            "description": "Tool output (only included when include.tool_calls.output is true)"
          },
          "duration_ms": {
            "type": "integer",
            "title": "Duration Ms",
            "description": "Execution time in milliseconds"
          },
          "iteration": {
            "type": "integer",
            "title": "Iteration",
            "description": "Iteration number (1-based) when this tool was called",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "tool",
          "input",
          "duration_ms"
        ],
        "title": "ReflectToolCall",
        "description": "A tool call made during reflect agent execution."
      },
      "ReflectTrace": {
        "properties": {
          "tool_calls": {
            "items": {
              "$ref": "#/components/schemas/ReflectToolCall"
            },
            "type": "array",
            "title": "Tool Calls",
            "description": "Tool calls made during reflection"
          },
          "llm_calls": {
            "items": {
              "$ref": "#/components/schemas/ReflectLLMCall"
            },
            "type": "array",
            "title": "Llm Calls",
            "description": "LLM calls made during reflection"
          }
        },
        "type": "object",
        "title": "ReflectTrace",
        "description": "Execution trace of LLM and tool calls during reflection."
      },
...
      "TokenUsage": {
        "properties": {
          "input_tokens": {
            "type": "integer",
            "title": "Input Tokens",
            "description": "Number of input/prompt tokens consumed",
            "default": 0
          },
          "output_tokens": {
            "type": "integer",
            "title": "Output Tokens",
            "description": "Number of output/completion tokens generated",
            "default": 0
          },
          "total_tokens": {
            "type": "integer",
            "title": "Total Tokens",
            "description": "Total tokens (input + output)",
            "default": 0
          }
        },
        "type": "object",
        "title": "TokenUsage",
        "description": "Token usage metrics for LLM calls.\n\nTracks input/output tokens for a single request to enable\nper-request cost tracking and monitoring.",
        "example": {
          "input_tokens": 1500,
          "output_tokens": 500,
          "total_tokens": 2000
        }
      },
      "ToolCallsIncludeOptions": {
        "properties": {
          "output": {
            "type": "boolean",
            "title": "Output",
            "description": "Include tool outputs in the trace. Set to false to only include inputs (smaller payload).",
            "default": true
          }
        },
        "type": "object",
        "title": "ToolCallsIncludeOptions",
        "description": "Options for including tool calls in reflect results."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}